Time: 7 ms (77.78%), Space: 41 MB (77.78%) - LeetHub

+++ 2327-number-of-people-aware-of-a-secret/2327-number-of-people-aware-of-a-secret.java
@@ -20,7 +20,7 @@ class Solution {
         // because it has only those number of people who remembers the
         // secret right now 
         for (int i = n - forget + 1; i <= n; ++i)
-            ans = (int)((ans + dp[i]) % mod);
+            ans = (ans + dp[i]) % mod;
         return ans;
     }
 }
